move functionality then deprecate speed and utils_graph modules for v2
